tests: distinct rpc contract address
