+++ conftest.py
@@ -231,6 +231,25 @@ def emitter_contract_address(client, accounts, EMITTER_CODE):
 
 
 @pytest.fixture()
+def rpc_emitter_contract_address(rpc_client, accounts, EMITTER_CODE):
+    txn_hash = rpc_client(
+        method="eth_sendTransaction",
+        params=[{
+            "from": accounts[0],
+            "data": EMITTER_CODE,
+        }],
+    )
+    txn_receipt = rpc_client(
+        method="eth_getTransactionReceipt",
+        params=[txn_hash],
+    )
+    assert txn_receipt is not None
+    assert txn_receipt['contractAddress']
+
+    return txn_receipt['contractAddress']
+
+
+@pytest.fixture()
 def call_emitter_contract(client, accounts, emitter_contract_address):
     def _call_emitter_contract(method_signature, arguments=None):
         if arguments is None:
@@ -247,6 +266,29 @@ def call_emitter_contract(client, accounts, emitter_contract_address):
         )
         return txn_hash
     return _call_emitter_contract
+
+
+@pytest.fixture()
+def call_emitter_method(rpc_client, rpc_emitter_contract_address, accounts):
+    def _call_emitter_method(method_signature, arguments=None):
+        if arguments is None:
+            arguments = []
+        function_sig = encode_data(sha3(method_signature)[:4])
+        data = function_sig + b''.join((strip_0x(encode_number(arg, 32)) for arg in arguments))
+        assert len(data) == 2 + 8 + 64 * len(arguments)
+
+        txn_hash = rpc_client(
+            method="eth_sendTransaction",
+            params=[{
+                '_from': accounts[0],
+                'to': rpc_emitter_contract_address,
+                'data': data,
+                'gas': 200000,
+            }],
+        )
+
+        return txn_hash
+    return _call_emitter_method
 
 
 class _Events(object):
